feat: common 모듈 작성 및 적용
