+++ service/common/src/main/java/org/codenbug/common/util/JwtUtil.java
@@ -1,4 +1,4 @@
-package org.codeNbug.queueserver.global.jwt;
+package org.codenbug.common.util;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
