+++ service/main-server/src/main/java/org/codeNbug/mainserver/domain/manager/controller/ManagerController.java
@@ -12,8 +12,8 @@ import org.codeNbug.mainserver.domain.manager.service.EventRegisterService;
 import org.codeNbug.mainserver.domain.manager.service.ManagerEventSearchService;
 import org.codeNbug.mainserver.domain.manager.service.ManagerPurchasesService;
 import org.codeNbug.mainserver.global.dto.RsData;
-import org.codeNbug.mainserver.global.security.annotation.RoleRequired;
-import org.codeNbug.mainserver.global.util.SecurityUtil;
+import org.codenbug.security.annotation.RoleRequired;
+import org.codenbug.security.util.SecurityUtil;
 import org.codenbug.user.constant.UserRole;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
