+++ service/main-server/src/main/java/org/codeNbug/mainserver/domain/notification/controller/NotificationController.java
@@ -4,8 +4,8 @@ import org.codeNbug.mainserver.domain.notification.dto.NotificationCreateRequest
 import org.codeNbug.mainserver.domain.notification.dto.NotificationDto;
 import org.codeNbug.mainserver.domain.notification.service.NotificationService;
 import org.codeNbug.mainserver.global.dto.RsData;
-import org.codeNbug.mainserver.global.security.annotation.RoleRequired;
-import org.codeNbug.mainserver.global.util.SecurityUtil;
+import org.codenbug.security.annotation.RoleRequired;
+import org.codenbug.security.util.SecurityUtil;
 import org.codenbug.user.constant.UserRole;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
