+++ service/main-server/src/main/java/org/codeNbug/mainserver/domain/user/controller/UserController.java
@@ -12,9 +12,9 @@ import org.codeNbug.mainserver.domain.user.service.UserService;
 import org.codeNbug.mainserver.global.Redis.service.TokenService;
 import org.codeNbug.mainserver.global.dto.RsData;
 import org.codeNbug.mainserver.global.exception.globalException.DuplicateEmailException;
-import org.codeNbug.mainserver.global.exception.security.AuthenticationFailedException;
-import org.codeNbug.mainserver.global.util.CookieUtil;
-import org.codeNbug.mainserver.global.util.SecurityUtil;
+import org.codenbug.common.util.CookieUtil;
+import org.codenbug.security.exception.AuthenticationFailedException;
+import org.codenbug.security.util.SecurityUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
