+++ service/main-server/src/main/java/org/codeNbug/mainserver/global/config/AppConfig.java
@@ -1,0 +1,9 @@
+package org.codeNbug.mainserver.global.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = {"org.codeNbug.mainserver", "org.codenbug.security", "org.codenbug.common"})
+public class AppConfig {
+}
