--- service/main-server/src/main/java/org/codeNbug/mainserver/global/exception/security/AuthenticationFailedException.java
@@ -1,10 +1,0 @@
-package org.codeNbug.mainserver.global.exception.security;
-
-/**
- * 인증 실패 예외
- */
-public class AuthenticationFailedException extends RuntimeException {
-    public AuthenticationFailedException(String message) {
-        super(message);
-    }
-}
