+++ service/queue-server/src/main/java/org/codeNbug/queueserver/global/GlobalConfig.java
@@ -1,9 +1,11 @@
 package org.codeNbug.queueserver.global;
 
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
 @EnableScheduling
+@ComponentScan(basePackages = {"org.codeNbug.queueserver", "org.codenbug.security", "org.codenbug.common.util"})
 public class GlobalConfig {
 }
