+++ service/queue-server/src/main/resources/application-prod.yml
@@ -6,3 +6,6 @@ db:
   username: root
   password: password
 
+cookie:
+  domain: localhost
+  secure: false  # dev 환경에서는 false, prod 환경에서는 true로 설정
