+++ service/queue-server/src/main/resources/application.yml
@@ -30,7 +30,10 @@ spring:
         default_batch_fetch_size: 100
         dialect: org.hibernate.dialect.MySQL8Dialect
     open-in-view: false
-
+# Cookie Configuration
+cookie:
+  domain: localhost
+  secure: false  # dev 환경에서는 false, prod 환경에서는 true로 설정
 jwt:
   expiration: 8640000
