+++ service/user/settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'user'
+
+include 'common'
