+++ service/user/src/main/java/org/codenbug/security/annotation/RoleRequired.java
@@ -1,5 +1,4 @@
-package org.codeNbug.mainserver.global.security.annotation;
-
+package org.codenbug.security.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,5 +10,5 @@ import org.codenbug.user.constant.UserRole;
 @Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface RoleRequired {
-    UserRole[] value() default {UserRole.USER};
-} 
+	UserRole[] value() default {UserRole.USER};
+}
