+++ service/user/src/main/java/org/codenbug/security/exception/AuthenticationFailedException.java
@@ -1,0 +1,10 @@
+package org.codenbug.security.exception;
+
+/**
+ * 인증 실패 예외
+ */
+public class AuthenticationFailedException extends RuntimeException {
+	public AuthenticationFailedException(String message) {
+		super(message);
+	}
+}
