+++ service/user/src/main/java/org/codenbug/security/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package org.codeNbug.mainserver.global.exception.security;
+package org.codenbug.security.exception;
 
 /**
  * 토큰이 유효하지 않을 때 발생하는 예외
@@ -7,4 +7,4 @@ public class InvalidTokenException extends RuntimeException {
     public InvalidTokenException(String message) {
         super(message);
     }
-} 
+}
