+++ service/user/src/main/java/org/codenbug/security/service/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package org.codeNbug.mainserver.global.security.service;
+package org.codenbug.security.service;
 
 import java.util.Collection;
 import java.util.Collections;
