+++ settings.gradle
@@ -5,3 +5,4 @@ include 'service:main-server'
 include 'service:queue-server'
 include 'service:message-dispatcher'
 include 'service:user'
+include 'service:common'
